cybersentinel-pro/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
python-nmap
beautifulsoup4
scapy
openai
psutil
cryptography
urllib3
